requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.7
    # via django
django==3.0.7
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.11.0
    # via -r requirements.in
pyparsing==2.4.7
    # via -r requirements.in
pytz==2019.1
    # via django
sqlparse==0.3.0
    # via django